coverity fixes...

+++ VRTestApp/FS.h
@@ -4,6 +4,6 @@ namespace zls
 {
 	namespace fs
 	{
-		void ReadFile(const char* fileName, char** content, int* len);
+		void ReadFile(const char* fileName, char** content, size_t* len);
 	}
 }
